feat : Lock 을 거는 서비스와 상위 클래스를 분리시킨다.
