+++ src/main/kotlin/com/hhplus/concert/interfaces/presentation/controller/BalanceController.kt
@@ -1,5 +1,6 @@
 package com.hhplus.concert.interfaces.presentation.controller
 
+import com.hhplus.concert.business.application.service.BalanceLockService
 import com.hhplus.concert.business.application.service.BalanceService
 import com.hhplus.concert.interfaces.presentation.request.BalanceRequest
 import com.hhplus.concert.interfaces.presentation.response.BalanceResponse
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RestController
 @RestController
 @RequestMapping("/api/v1/balance")
 class BalanceController(
+    private val balanceLockService: BalanceLockService,
     private val balanceService: BalanceService,
 ) {
     // 잔액을 충전한다.
@@ -22,7 +24,7 @@ class BalanceController(
         @RequestBody rechargeRequest: BalanceRequest.Recharge,
     ): BalanceResponse.Detail =
         BalanceResponse.Detail.from(
-            balanceService.recharge(userId, rechargeRequest.amount),
+            balanceLockService.recharge(userId, rechargeRequest.amount),
         )
 
     // 잔액을 조회한다.
